# api/openapi.yaml
openapi: 3.1.0
info:
  title: capp-backend API
  description: |
    Multi-cluster API gateway for managing Capp (container-app) resources on
    Kubernetes / OpenShift clusters.

    ## Authentication

    The backend supports five auth modes configured server-side:

    - **passthrough** — supply a raw Kubernetes `Bearer <token>` on every request.
    - **jwt** — call `/api/v1/auth/login` with `cluster` + `token` to exchange a
      cluster token for a short-lived JWT; refresh with `/api/v1/auth/refresh`.
    - **static** — use a fixed API key configured on the server (dev / CI only).
    - **dex** — call `/api/v1/auth/login` with `username` + `password`; Dex handles OIDC.
    - **openshift** — use `/api/v1/auth/openshift/authorize` to get an OAuth URL,
      then POST the code to `/api/v1/auth/openshift/callback`.

    Use `GET /api/v1/auth/mode` to discover which mode the server is running.
    All resource endpoints require the `Authorization: Bearer <token>` header.
  version: "0.1.0"
  contact:
    name: dana-team
    url: https://github.com/dana-team/capp-backend
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        In **passthrough** mode supply the raw Kubernetes ServiceAccount token.
        In **jwt** / **dex** / **openshift** modes supply the JWT returned by the
        login or callback endpoint.
        In **static** mode supply one of the configured API keys.
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: resource "my-capp" not found
            status:
              type: integer
              example: 404

    # ── Auth ─────────────────────────────────────────────────────────────────────
...
    NamespaceItem:
      type: object
      properties:
        name:
          type: string
          example: default
        status:
          type: string
          example: Active
    NamespaceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NamespaceItem'
        canCreate:
          type: boolean
          description: Whether the authenticated user has permission to create new namespaces in this cluster.
    NamespaceQuota:
      type: object
      description: Resource quota limits to enforce on the namespace.
      properties:
        cpu:
          type: string
          description: CPU limit quota (Kubernetes quantity format).
          example: "4"
        memory:
          type: string
          description: Memory limit quota (Kubernetes quantity format).
          example: 8Gi
        pods:
          type: integer
          description: Maximum number of pods allowed in the namespace.
          example: 50
    CreateNamespaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Name of the namespace to create.
          example: my-namespace
        users:
          type: array
          items:
            type: string
          description: |
            Users to bind to the `capp-user` ClusterRole via a RoleBinding.
            If omitted, no RoleBinding is created.
          example: [alice, bob]
        quota:
          $ref: '#/components/schemas/NamespaceQuota'
    UpdateNamespaceRequest:
      type: object
      description: |
        Replaces the namespace's ResourceQuota and RoleBinding. Omit a field to
        leave that resource unchanged. Quota decreases are rejected when current
        usage exceeds the requested new limit.
      properties:
        users:
          type: array
          items:
            type: string
          description: |
            Full replacement list of users for the `capp-user` RoleBinding.
            Creates the RoleBinding if it does not exist.
          example: [alice, bob]
        quota:
          $ref: '#/components/schemas/NamespaceQuota'
    PatchNamespaceRequest:
      type: object
      description: Adds new users to the namespace's existing `capp-user` RoleBinding without removing current members.
      properties:
        users:
          type: array
          items:
            type: string
          description: Users to add. Existing users are deduplicated; no user is removed.
          example: [charlie]

    # ── ConfigMaps ────────────────────────────────────────────────────────────────
...
  parameters:
    clusterParam:
      name: cluster
      in: path
      required: true
      schema:
        type: string
      description: Cluster name as configured on the backend.
      example: production
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The caller does not have permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: A resource with this name already exists.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []
paths:
...
  /api/v1/clusters/{cluster}/namespaces:
    get:
      tags: [namespaces]
      summary: List namespaces in a cluster
      description: |
        Returns Kubernetes namespaces (or OpenShift Projects) the authenticated
        user has access to. Only namespaces labeled `dana.io/capp-ns=true` are
        returned. On vanilla Kubernetes each namespace is checked via
        SelfSubjectAccessReview; on OpenShift the Projects API handles access
        control natively.
      parameters:
        - $ref: '#/components/parameters/clusterParam'
      responses:
        "200":
          description: Namespace list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceList'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "503":
          description: Cluster is currently unhealthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [ namespaces ]
      summary: Create a namespace in a cluster
      description: |
        Creates a new Kubernetes namespace (or OpenShift Project) in the specified
        cluster. The namespace is automatically labeled `dana.io/capp-ns=true` so
        it is visible in the namespace list endpoint.
      parameters:
        - $ref: '#/components/parameters/clusterParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNamespaceRequest'
      responses:
        "201":
          description: Namespace created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceItem'
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
        "500":
          $ref: '#/components/responses/InternalError'
  /api/v1/clusters/{cluster}/namespaces/{name}:
    put:
      tags: [namespaces]
      summary: Replace namespace quota and role binding
      description: |
        Full replacement of the namespace's ResourceQuota and `capp-user` RoleBinding.
        Omit `quota` or `users` to leave that resource unchanged.
        Quota decreases are rejected when current usage exceeds the new limit.
      parameters:
        - $ref: '#/components/parameters/clusterParam'
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Namespace name.
          example: my-namespace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNamespaceRequest'
      responses:
        "200":
          description: Namespace updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceItem'
        "400":
          description: Invalid request body or quota decrease rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [namespaces]
      summary: Add users to namespace role binding
      description: |
        Appends new users to the namespace's existing `capp-user` RoleBinding.
        Existing members are preserved; duplicate users are silently ignored.
        The RoleBinding must already exist (created via POST or PUT).
      parameters:
        - $ref: '#/components/parameters/clusterParam'
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Namespace name.
          example: my-namespace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchNamespaceRequest'
      responses:
        "200":
          description: Role binding updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamespaceItem'
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalError'

  # ── ConfigMaps ─────────────────────────────────────────────────────────────────
...